requirements.txt
alembic==1.14.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
black==26.3.1
certifi==2026.4.22
charset-normalizer==3.4.7
choreographer==1.2.1
click==8.3.3
coverage==7.13.5
exceptiongroup==1.3.1
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
importlib_metadata==9.0.0
iniconfig==2.3.0
kaleido==1.2.0
logistro==2.0.1
mypy_extensions==1.1.0
narwhals==2.20.0
numpy==2.2.6
orjson==3.11.8
packaging==26.1
pandas==2.3.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pathspec==1.1.0
platformdirs==4.9.6
plotly==6.7.0
pluggy==1.6.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
pytokens==0.4.1
pytz==2026.1.post1
requests==2.33.1
ruff==0.15.11
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.46.0
zipp==3.23.1
